Moving the CurrentGameEndpoint and FeaturedGameEndpoint to a common SpectatorEndpoint. I hope it's okay like this.
