+++ RiotSharp/Interfaces/IRiotApi.cs
@@ -1,6 +1,4 @@
 ﻿using RiotSharp.ChampionEndpoint;
-using RiotSharp.CurrentGameEndpoint;
-using RiotSharp.FeaturedGamesEndpoint;
 using RiotSharp.GameEndpoint;
 using RiotSharp.LeagueEndpoint;
 using RiotSharp.MatchEndpoint;
@@ -12,6 +10,7 @@ using System.Threading.Tasks;
 using System;
 using RiotSharp.MatchListEndpoint;
 using RiotSharp.Misc;
+using RiotSharp.SpectatorEndpoint;
 
 namespace RiotSharp.Interfaces
 {
