+++ RiotSharp/SpectatorEndpoint/CurrentGame/BannedChampion.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace RiotSharp.CurrentGameEndpoint
+namespace RiotSharp.SpectatorEndpoint.CurrentGame
 {
     /// <summary>
     /// Class representing a BannedChampion in the API.
